refactor(storybook): extract ErrorAndValidationWorkbench to its own file
Exporting the workbench component directly from LayoutPatterns.stories.tsx
caused Storybook's CSF auto-detection to treat it as an implicit story
(any named export from a .stories.tsx file is picked up), creating a
duplicate "Error And Validation Workbench" entry in the sidebar alongside
the real "UX Error and Validation" story. Moving it to a plain .tsx module
lets it stay importable by both LayoutPatterns and the new Code Editors
pattern pages without being auto-registered as a story.
